b/375512690 Merge Ssh, Rdp into Session namespace
