+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/ToolWindows/Session/TestConnectCommands.cs
@@ -27,7 +27,6 @@ using Google.Solutions.IapDesktop.Extensions.Session.Protocol;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Rdp;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Ssh;
 using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Session;
-using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Ssh;
 using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Ssh;
 using Moq;
