+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/ToolWindows/Session/TestRdpView.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright 2024 Google LLC
+// Copyright 2020 Google LLC
 //
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
@@ -21,22 +21,17 @@
 
 using Google.Solutions.Apis.Auth;
 using Google.Solutions.Apis.Compute;
-using Google.Solutions.Apis.Crm;
 using Google.Solutions.Apis.Locator;
+using Google.Solutions.Common.Security;
 using Google.Solutions.IapDesktop.Application.Theme;
 using Google.Solutions.IapDesktop.Application.Windows;
-using Google.Solutions.IapDesktop.Application.Windows.Dialog;
 using Google.Solutions.IapDesktop.Core.ClientModel.Transport;
 using Google.Solutions.IapDesktop.Core.ObjectModel;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Rdp;
-using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Ssh;
-using Google.Solutions.IapDesktop.Extensions.Session.Settings;
 using Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.Rdp;
 using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Session;
-using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Ssh;
 using Google.Solutions.Mvvm.Binding;
-using Google.Solutions.Ssh;
-using Google.Solutions.Ssh.Cryptography;
+using Google.Solutions.Terminal.Controls;
 using Google.Solutions.Testing.Apis.Integration;
 using Google.Solutions.Testing.Application.ObjectModel;
 using Google.Solutions.Testing.Application.Views;
@@ -45,37 +40,60 @@ using NUnit.Framework;
 using System;
 using System.Drawing;
 using System.Net;
-using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.Ssh
+namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.Session
 {
     [TestFixture]
     [UsesCloudResources]
-    public class TestSshView : WindowTestFixtureBase
+    [RequiresRdp]
+    public class TestRdpView : WindowTestFixtureBase
     {
+        private static async Task<RdpCredential> GenerateRdpCredentialAsync(
+            InstanceLocator instance)
+        {
+            var username = "test" + Guid.NewGuid().ToString().Substring(0, 4);
+            var credentialAdapter = new WindowsCredentialGenerator(
+                new ComputeEngineClient(
+                    ComputeEngineClient.CreateEndpoint(),
+                    TestProject.AdminAuthorization,
+                    TestProject.UserAgent));
+
+            var credential = await credentialAdapter
+                .CreateWindowsCredentialsAsync(
+                    instance,
+                    username,
+                    UserFlags.AddToAdministrators,
+                    TimeSpan.FromSeconds(60),
+                    CancellationToken.None)
+                .ConfigureAwait(true);
+
+            return new RdpCredential(
+                credential.UserName,
+                credential.Domain,
+                credential.SecurePassword);
+        }
+
+        //
+        // Use a larger machine type as all this RDP'ing consumes a fair
+        // amount of memory.
+        //
+        private const string MachineTypeForRdp = "n1-highmem-2";
+
         private static readonly InstanceLocator SampleLocator =
             new InstanceLocator("project", "zone", "instance");
 
         private IServiceProvider CreateServiceProvider(IAuthorization authorization)
         {
             var registry = new ServiceRegistry(this.ServiceRegistry);
-            registry.AddTransient<SshView>();
-            registry.AddTransient<SshViewModel>();
+            registry.AddTransient<RdpView>();
+            registry.AddTransient<RdpViewModel>();
             registry.AddMock<IBindingContext>();
             registry.AddMock<IToolWindowTheme>();
-            registry.AddMock<IInputDialog>();
             registry.AddTransient<IToolWindowHost, ToolWindowHost>();
             registry.AddSingleton(authorization);
-            registry.AddSingleton<SessionFactory>();
-
-            var settingsRepository = registry.AddMock<ITerminalSettingsRepository>();
-            settingsRepository
-                .Setup(r => r.GetSettings())
-                .Returns(TerminalSettings.Default);
-            
             return registry;
         }
 
@@ -85,34 +103,6 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.Ssh
             transport.SetupGet(t => t.Protocol).Returns(RdpProtocol.Protocol);
             transport.SetupGet(t => t.Endpoint).Returns(endpoint);
             return transport.Object;
-        }
-
-        private static async Task<ISshCredential> CrateSshCredential(
-            InstanceLocator instance,
-            IAuthorization authorization,
-            SshKeyType keyType)
-        {
-            var keyAdapter = new PlatformCredentialFactory(
-                authorization,
-                new ComputeEngineClient(
-                    ComputeEngineClient.CreateEndpoint(),
-                    authorization,
-                    TestProject.UserAgent),
-                new ResourceManagerClient(
-                    ResourceManagerClient.CreateEndpoint(),
-                    authorization,
-                    TestProject.UserAgent),
-                new Mock<IOsLoginProfile>().Object);
-
-            return await keyAdapter
-                .CreateCredentialAsync(
-                    instance,
-                    AsymmetricKeySigner.CreateEphemeral(keyType),
-                    TimeSpan.FromMinutes(10),
-                    null,
-                    KeyAuthorizationMethods.InstanceMetadata,
-                    CancellationToken.None)
-                .ConfigureAwait(true);
         }
 
         //---------------------------------------------------------------------
@@ -127,20 +117,24 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.Ssh
             var transport = CreateTransportForEndpoint(unboundEndpoint);
 
             var serviceProvider = CreateServiceProvider(await auth);
-            var factory = serviceProvider.GetService<SessionFactory>();
-
+            var broker = new SessionFactory(
+                serviceProvider.GetService<IMainWindow>(),
+                serviceProvider.GetService<ISessionBroker>(),
+                serviceProvider.GetService<IToolWindowHost>(),
+                serviceProvider.GetService<IJobService>());
             await AssertRaisesEventAsync<SessionAbortedEvent>(
-                () => factory.ConnectSshSession(
+                () => broker.ConnectRdpSession(
                     SampleLocator,
                     transport,
-                    new SshParameters()
+                    new RdpParameters()
                     {
                         ConnectionTimeout = TimeSpan.FromSeconds(5)
                     },
-                    new StaticPasswordCredential("user", string.Empty)))
+                    RdpCredential.Empty))
                 .ConfigureAwait(true);
 
-            Assert.IsInstanceOf(typeof(SocketException), this.ExceptionShown);
+            Assert.IsInstanceOf(typeof(RdpDisconnectedException), this.ExceptionShown);
+            Assert.AreEqual(516, ((RdpDisconnectedException)this.ExceptionShown!).DisconnectReason);
         }
 
         //---------------------------------------------------------------------
@@ -149,60 +143,76 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.ToolWindows.Ssh
 
         [Test]
         public async Task Connect_WhenCredentialsInvalid_ThenErrorIsShownAndWindowIsClosed(
-            [LinuxInstance] ResourceTask<InstanceLocator> testInstance,
+            [WindowsInstance(MachineType = MachineTypeForRdp)] ResourceTask<InstanceLocator> testInstance,
             [Credential(Role = PredefinedRole.IapTunnelUser)] ResourceTask<IAuthorization> auth)
         {
+            var serviceProvider = CreateServiceProvider(await auth);
             var instance = await testInstance;
 
-            using (var tunnel = IapTransport.CreateSshTransport(
+            using (var tunnel = IapTransport.CreateRdpTransport(
                 instance,
                 await auth))
             {
-                var serviceProvider = CreateServiceProvider(await auth);
-                var factory = serviceProvider.GetService<SessionFactory>();
+                var rdpCredential = new RdpCredential(
+                    "wrong",
+                    null,
+                    SecureStringExtensions.FromClearText("wrong"));
+                var rdpParameters = new RdpParameters()
+                {
+                    AuthenticationLevel = RdpAuthenticationLevel.NoServerAuthentication,
+                    UserAuthenticationBehavior = RdpAutomaticLogon.LegacyAbortOnFailure
+                };
+
+                var broker = new SessionFactory(
+                    serviceProvider.GetService<IMainWindow>(),
+                    serviceProvider.GetService<ISessionBroker>(),
+                    serviceProvider.GetService<IToolWindowHost>(),
+                    serviceProvider.GetService<IJobService>());
 
                 await AssertRaisesEventAsync<SessionAbortedEvent>(
-                    () => factory.ConnectSshSession(
+                    () => broker.ConnectRdpSession(
                         instance,
                         tunnel,
-                        new SshParameters(),
-                        new StaticPasswordCredential("user", string.Empty)))
+                        rdpParameters,
+                        rdpCredential))
                     .ConfigureAwait(true);
 
                 Assert.IsNotNull(this.ExceptionShown);
-                Assert.IsInstanceOf(typeof(UnsupportedAuthenticationMethodException), this.ExceptionShown);
+                Assert.IsInstanceOf(typeof(RdpDisconnectedException), this.ExceptionShown);
+                Assert.AreEqual(2055, ((RdpDisconnectedException)this.ExceptionShown!).DisconnectReason);
             }
         }
 
         [Test]
         public async Task Connect_WhenWindowChangedToFloating_ThenConnectionSurvives(
-            [LinuxInstance] ResourceTask<InstanceLocator> testInstance,
-            [Credential(Roles = new[] {
-                PredefinedRole.IapTunnelUser,
-                PredefinedRole.ComputeInstanceAdminV1
-            })] ResourceTask<IAuthorization> auth)
+            [WindowsInstance(MachineType = MachineTypeForRdp)] ResourceTask<InstanceLocator> testInstance,
+            [Credential(Role = PredefinedRole.IapTunnelUser)] ResourceTask<IAuthorization> auth)
         {
+            var serviceProvider = CreateServiceProvider(await auth);
             var instance = await testInstance;
+            var rdpCredential = await
+                GenerateRdpCredentialAsync(instance)
+                .ConfigureAwait(true);
 
-            using (var tunnel = IapTransport.CreateSshTransport(
+            using (var tunnel = IapTransport.CreateRdpTransport(
                 instance,
                 await auth))
             {
-                var serviceProvider = CreateServiceProvider(await auth);
-                var factory = serviceProvider.GetService<SessionFactory>();
+                var rdpParameters = new RdpParameters();
 
-                var credential = await CrateSshCredential(
-                    instance,
-                    await auth,
-                    SshKeyType.Rsa3072);
+                var broker = new SessionFactory(
+                    serviceProvider.GetService<IMainWindow>(),
+                    serviceProvider.GetService<ISessionBroker>(),
+                    serviceProvider.GetService<IToolWindowHost>(),
+                    serviceProvider.GetService<IJobService>());
 
-                SshView? session = null;
+                RdpView? session = null;
                 await AssertRaisesEventAsync<SessionStartedEvent>(
-                    () => session = (SshView)factory.ConnectSshSession(
+                    () => session = (RdpView)broker.ConnectRdpSession(
                         instance,
                         tunnel,
-                        new SshParameters(),
-                        credential))
+                        rdpParameters,
+                        rdpCredential))
                     .ConfigureAwait(true);
 
                 Assert.IsNotNull(session);
