+++ sources/Google.Solutions.IapDesktop.Extensions.Session/ToolWindows/Session/ConnectCommands.cs
@@ -24,7 +24,6 @@ using Google.Solutions.IapDesktop.Core.ObjectModel;
 using Google.Solutions.IapDesktop.Core.ProjectModel;
 using Google.Solutions.IapDesktop.Extensions.Session.Properties;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol;
-using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Ssh;
 using Google.Solutions.Mvvm.Binding.Commands;
 using System.Threading;
 using System.Threading.Tasks;
