+++ sources/Google.Solutions.IapDesktop.Extensions.Session/ToolWindows/Session/ConnectInstanceCommand.cs
@@ -22,8 +22,6 @@
 using Google.Solutions.IapDesktop.Application.Windows;
 using Google.Solutions.IapDesktop.Core.ProjectModel;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol;
-using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Rdp;
-using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Ssh;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
