+++ sources/Google.Solutions.IapDesktop.Extensions.Session/ToolWindows/Session/ConnectRdpUrlCommand.cs
@@ -22,7 +22,6 @@
 using Google.Solutions.IapDesktop.Application.Data;
 using Google.Solutions.IapDesktop.Application.Windows;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol;
-using Google.Solutions.IapDesktop.Extensions.Session.ToolWindows.Rdp;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
